+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 * nostr: add `EventBuilder::git_issue` ([Yuki Kishimoto])
 * nostr: add `EventBuilder::git_patch` ([Yuki Kishimoto])
 * nostr: add `Tag::reference` constructor ([Yuki Kishimoto])
+* nostr: add `nip59::make_seal` function ([Yuki Kishimoto])
 * database: add `Backend::is_persistent` method ([Yuki Kishimoto])
 * database: add `Events` struct ([Yuki Kishimoto])
 * relay-builder: add `LocalRelay` and `RelayBuilder` ([Yuki Kishimoto])